+++ jarboot-server/src/main/java/com/mz/jarboot/controller/SettingController.java
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @Api(tags="系统配置")
-@RequestMapping(value = "/api/jarboot-setting")
+@RequestMapping(value = "/api/jarboot/setting")
 @Controller
 @Permission
 public class SettingController {