+++ jarboot-server/src/main/java/com/mz/jarboot/service/impl/UploadFileServiceImpl.java
@@ -133,9 +133,8 @@ public class UploadFileServiceImpl implements UploadFileService {
         File dest = FileUtils.getFile(destPath);
         //开始复制前要不要先备份，以便失败后还原？文件量、体积巨大如何处理？为了性能先不做考虑
         try {
-            String[] ex = new String[]{"jar"};
             //先复制jar文件
-            Collection<File> jarFiles = FileUtils.listFiles(dir, ex, false);
+            Collection<File> jarFiles = FileUtils.listFiles(dir, CommonConst.JAR_FILE_EXT, false);
             for (File jar : jarFiles) {
                 FileUtils.copyFileToDirectory(jar, dest, true);
             }
@@ -144,7 +143,7 @@ public class UploadFileServiceImpl implements UploadFileService {
             //检测多个jar文件时有没有配置启动的jar文件
             ServerSettingDTO setting = PropertyFileUtils.getServerSetting(server);
             if (StringUtils.isEmpty(setting.getJar())) {
-                boolean bo = FileUtils.listFiles(dest, ex, false).size() > 1;
+                boolean bo = FileUtils.listFiles(dest, CommonConst.JAR_FILE_EXT, false).size() > 1;
                 if (bo) {
                     String msg = String.format("在服务%s目录找到了多个jar文件，请设置启动的jar文件！", server);
                     WebSocketManager.getInstance().notice(msg, NoticeEnum.WARN);