+++ jarboot-server/src/main/java/com/mz/jarboot/utils/SettingUtils.java
@@ -5,6 +5,7 @@ import com.mz.jarboot.common.OSUtils;
 import com.mz.jarboot.common.ResultCodeConst;
 import com.mz.jarboot.constant.CommonConst;
 import com.mz.jarboot.common.MzException;
+import com.mz.jarboot.constant.SettingPropConst;
 import com.mz.jarboot.dto.GlobalSettingDTO;
 import com.mz.jarboot.dto.ServerSettingDTO;
 import com.mz.jarboot.event.ApplicationContextUtils;
@@ -58,10 +59,10 @@ public class SettingUtils {
         File conf = new File(JARBOOT_CONF);
         Properties properties = (conf.exists() && conf.isFile() && conf.canRead()) ?
                 PropertyFileUtils.getProperties(conf) : new Properties();
-        globalSetting.setServicesPath(properties.getProperty(ROOT_DIR_KEY, ""));
-        globalSetting.setDefaultJvmArg(properties.getProperty(DEFAULT_JVM_OPTS_KEY, ""));
-        String s = properties.getProperty(ENABLE_AUTO_START_KEY, "false");
-        boolean servicesAutoStart = StringUtils.equalsIgnoreCase("true", s);
+        globalSetting.setServicesPath(properties.getProperty(ROOT_DIR_KEY, StringUtils.EMPTY));
+        globalSetting.setDefaultJvmArg(properties.getProperty(DEFAULT_JVM_OPTS_KEY, StringUtils.EMPTY));
+        String s = properties.getProperty(ENABLE_AUTO_START_KEY, SettingPropConst.VALUE_FALSE);
+        boolean servicesAutoStart = StringUtils.equalsIgnoreCase(SettingPropConst.VALUE_TRUE, s);
         globalSetting.setServicesAutoStart(servicesAutoStart);
     }
 
@@ -144,8 +145,9 @@ public class SettingUtils {
             WebSocketManager.getInstance().notice("未找到服务" + server + "的可执行jar包路径", NoticeEnum.WARN);
         }
         if (StringUtils.isNotEmpty(setting.getJar())) {
-            //配置制定了jar文件
-            File jar = FileUtils.getFile(dir, setting.getJar());
+            //配置了jar文件，判定是否绝对路径
+            Path path = Paths.get(setting.getJar());
+            File jar = path.isAbsolute() ? path.toFile() : FileUtils.getFile(dir, setting.getJar());
             if (jar.exists() && jar.isFile()) {
                 return jar.getPath();
             } else {
@@ -155,23 +157,22 @@ public class SettingUtils {
         }
 
         //未指定时
-        String[] extensions = {"jar"};
-        Collection<File> jarList = FileUtils.listFiles(dir, extensions, false);
+        Collection<File> jarList = FileUtils.listFiles(dir, CommonConst.JAR_FILE_EXT, false);
         if (org.apache.commons.collections.CollectionUtils.isEmpty(jarList)) {
             logger.error("在{}未找到{}服务的jar包", server, dir.getPath());
             WebSocketManager.getInstance().notice("未找到服务" + server + "的可执行jar包", NoticeEnum.ERROR);
-            return "";
+            return StringUtils.EMPTY;
         }
         if (jarList.size() > 1) {
             String msg = String.format("在服务%s目录找到了多个jar文件，请设置启动的jar文件！", server);
             WebSocketManager.getInstance().notice(msg, NoticeEnum.WARN);
-            return "";
+            return StringUtils.EMPTY;
         }
         if (jarList.iterator().hasNext()) {
             File jarFile = jarList.iterator().next();
             return jarFile.getPath();
         }
-        return "";
+        return StringUtils.EMPTY;
     }
 
     public static String getServerPath(String server) {
@@ -189,7 +190,7 @@ public class SettingUtils {
 
     public static String getJvm(String server, String file) {
         if (StringUtils.isBlank(file)) {
-            file = CommonConst.DEFAULT_JVM_FILE;
+            file = SettingPropConst.DEFAULT_VM_FILE;
         }
         Path path = Paths.get(file);
         if (!path.isAbsolute()) {